+++ packages/client/src/pages/Forum/ForumTopic/ForumMessage/ForumMessage.tsx
@@ -3,7 +3,7 @@ import './ForumMessage.css';
 import cn from 'classnames';
 import { type FC, memo, useCallback, useState } from 'react';
 
-import { forumAPI } from '../../../../api/forumAPI';
+import { forumAPI } from '../../../../api/forumAPI/forumAPI';
 import { Button } from '../../../../components/Button';
 import { ButtonVariant } from '../../../../components/Button/data';
 import { Dropdown } from '../../../../components/Dropdown';
