+++ packages/client/src/pages/Forum/ForumTopic/ForumTopic.tsx
@@ -4,7 +4,7 @@ import cn from 'classnames';
 import { type FC, useCallback, useState } from 'react';
 import { useLoaderData, useParams } from 'react-router-dom';
 
-import { forumAPI } from '../../../api/forumAPI';
+import { forumAPI } from '../../../api/forumAPI/forumAPI';
 import { Breadcrumbs } from '../../../components/Breadcrumbs';
 import { BreadcrumbsVariant } from '../../../components/Breadcrumbs/data';
 import { Button } from '../../../components/Button';
