+++ packages/client/src/store/features/auth/authThunks.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
-import { type SigninRequestData, type SignupRequestData, authAPI } from '../../../api/authAPI';
+import { type SigninRequestData, type SignupRequestData, authAPI } from '../../../api/authAPI/authAPI';
 
 /**
  * Thunks - вызывают методы апи и полученные данные автоматически записывают в хранилище.
