+++ packages/server/api/services/ForumTopic.ts
@@ -13,56 +13,50 @@ export const forumTopicRoute = Router()
   .use(express.urlencoded({ extended: true }))
   .use('/', checkAuthMiddleware)
   .get('/', (req: Request, res: Response, next) => {
-    ForumTopic.findAll({ where: { section_id: req.query.section_id }})
+    ForumTopic.findAll({ where: { section_id: req.query.section_id } })
       .then((topics: ForumTopic[]) => res.status(200).json(topics))
       .catch(next);
   })
   .get('/:id', (req: Request, res: Response, next) => {
     ForumTopic.findByPk(req.params.id, {
-      include: [
-        { model: ForumSection },
-        { model: ForumMessage,
-          include: [{ model: User }],
-        },
-        { model: User }],
-      order: [
-        [ Sequelize.col('messages.created_at'), 'ASC' ]
-      ]
+      include: [{ model: ForumSection }, { model: ForumMessage, include: [{ model: User }] }, { model: User }],
+      order: [[Sequelize.col('messages.created_at'), 'ASC']],
     })
       .then(throwIf(r => !r, res, 400, 'Тема не найдена'))
       .then(topic => res.status(200).json(topic))
       .catch(next);
   })
   .post('/', (req: Request, res: Response, next) => {
-    if (res.locals.user && res.locals.user.id ) {
+    if (res.locals.user && res.locals.user.id) {
       req.body.user_id = res.locals.user.id;
-      ForumTopic.create(req.body).then(topic => res.status(201).send({ id: topic.id }))
-     .catch(next);
+      ForumTopic.create(req.body)
+        .then(topic => res.status(201).send({ id: topic.id }))
+        .catch(next);
     } else {
       res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
     }
   })
   .put('/:id', (req: Request, res: Response, next) => {
     if (res.locals.user && res.locals.user.id) {
-    ForumTopic.update(req.body, { where: { id: req.params.id, user_id: res.locals.user.id }, returning: true })
-      .then(result => {
-        const [count, topics] = result;
-        if (count === 0) {
-          throw Error('Топик не найден');
-        }
-        res.status(200).json(topics[0]);
-      })
-      .catch(next);
+      ForumTopic.update(req.body, { where: { id: req.params.id, user_id: res.locals.user.id }, returning: true })
+        .then(result => {
+          const [count, topics] = result;
+          if (count === 0) {
+            throw Error('Топик не найден');
+          }
+          res.status(200).json(topics[0]);
+        })
+        .catch(next);
     } else {
       res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
     }
   })
   .delete('/:id', (req: Request, res: Response, next) => {
     if (res.locals.user && res.locals.user.id) {
-    ForumTopic.destroy({ where: { id: req.params.id, user_id: res.locals.user.id } })
-      .then(throwIf(r => !r, res, 400, 'Тема не найдена'))
-      .then(topic => res.status(201).json(topic))
-      .catch(next);
+      ForumTopic.destroy({ where: { id: req.params.id, user_id: res.locals.user.id } })
+        .then(throwIf(r => !r, res, 400, 'Тема не найдена'))
+        .then(topic => res.status(201).json(topic))
+        .catch(next);
     } else {
       res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
     }
