+++ packages/server/tsconfig.json
@@ -20,21 +20,12 @@
     "importsNotUsedAsValues": "error",
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
-    "lib": [
-      "DOM",
-      "es2019",
-      "esnext.asynciterable"
-    ],
-    "types": [
-      "node",
-      "jest"
-    ],
+    "lib": ["DOM", "es2019", "esnext.asynciterable"],
+    "types": ["node", "jest"],
     "baseUrl": ".",
     "paths": {
-      "*": [
-        "types/*"
-      ]
+      "*": ["types/*"]
     },
     "outDir": "dist"
-  },
+  }
 }
